fixed some para sql files (#367)
Current workflow for rapid onset flooding probability looks at the first instance of ROF among all the members. I updated the logic to track ROF conditions for all instances and tracked which days experienced ROF. I then used that to calculate the probability. I also updated some of the workflow to use more pandas as well as update some of the short range version variables to better align with the NWM file language of model initialization and model output.