Fixes regex matching bug (#473)
The replace and route execution of the initialize_pipeline lambda began failing in TI a few days ago. I'm guessing I must have had a hard-coded in-place change directly on TI from back when I was developing and testing the replace and route code. Again as my guess, it's likely that TI was just redeployed a couple of days ago with the rnr code that had been recently merged into it.

Anyhow, this small change will fix the issue.