added sleep after getting server (#518)
When the publish service lambda would use the gis.content.search functionality, the response would sometimes be an empty list. This was often seen in the lambda in the non-active AWS region. I wonder if this has to do with the DNS routing or something. Either way, I updated to the code to instead grab the portal items from the service and then use the gis.content.get method using the portal item ids.

This is a more robust method for 2 reasons

1. We ensure that we are updating the correct portal items that are associated with the service.
2. We are also now updating both the map service and the feature service portal items where we were only updating the map service portal item before

---------

replaced content search with content get based on service portal item ids