updated the hand_stage field to be fim_stage (#526)
In PR #519, the 14day inundation extent services fields were changed from fim_stage (the correct new field) back to the hand_stage (the deprecated old field). This got past me when I reviewed that PR and didn't catch it. This PR fixes that issue so that the service points to the correct field and renders correctly.