added empty rows so fimpact counties for hi and prvi (#525)
The Ops monitor currently checks the fimpact layers (counties specifically) for the reference time for Hawaii and Puerto Rico inundation services. However, these layers will be empty if there is no inundation. This PR will add an empty row to the county fimpact layers for these domains so that the Ops monitor can get the correct reference time.