Rio Grande FIM Removal (#534)
This PR removes the Rio Grande FIM from public services. Eventually we will want to remove all non-CONUS reaches out of all services but this is immediate action needed.

In addition to these code changes, I also updated the derived.channels_conus table with a new "public_fim" column. This is a boolean column which indicates if certain reached can be public or not.

-------------

* updated summary sql to include selecting feature where public fim is true
I also updated the channels conus table with a new public fim column for this

* updated fim_domain to be public_fim_domain for more clarity