Implements coastal fim processing by tile, not huc
Refs #538
The above ticket was created to address the issue of certain HUCs not
rendering in the coastal fim services. The problem was that VRTs render
in layered order, and the HUC dems were overlapping each other and
thus covering up actual fim data. This is because the HUC-based dems
are still rectangular (i.e. the bbox of the HUC) and not the shape of
the HUCs themselves. It's true that all of the non-HUC grid cells are
marked as no-data, but it's these no-data cells that are getting layer-
ordering priority and masking the actual underlying HUC data cells.
The solution was to subset the coastal DEMs to non-overlapping
rectangular, tile-based DEMs for processing. These can then be
rendered without any overlaps. It also results in many fewer rasters
for both processing and rendering, thus speeding up the process.
I also discovered while testing that the VRT pointing to the underlying
MRF files was rendering much slower than if I pointed the VRT to the
TIF files instead. So that change is also incorporated here.
Lastly, I converted the viz_schism_fim_proceesing lambda processing code
to operate completely on MemoryFile objects, rather than writing to
disk. This also speeds up the processing a bit.