fixed regex bug for getting forecast timestep from file name (#541)
This PR fixes a regex issue when trying to grab the forecast timestep from the file name. For some reason the wildcard with a digit (\d*) was not grabbing the number correctly. I updated the regex (\d{1,9}) to look for a certain amount of digits. The numbers used were a bit arbitrary. I just wanted it to be small enough and big enough to work and then added some buffer just in case future files would need it.