New fields added and old fields refactored (#560)
This commit changes 36 files - 18 of which are product SQL files and the other 18 are the MAPX files associated with those same products. In every case, the following was done if/where applicable:

 - New fields were created alongside any "pre-existing fields that reported a time as an hour" that convert the associated hour into an actual datetime string. All of these new fields were named appropriately to end in "_time" (e.g. "high_water_arrival_time" or "below_bank_return_time").

 - All pre-existing fields that reported a "time" as an hour were renamed to end in "_hour" if they did not already do so. In some cases these pre-existing fields ended in "_time" (even though they stored an hour) instead of "_hour" and thus "_time" was just replaced with "_hour" to avoid conflicting with the names of the new fields mentioned above. In other cases, the pre-existing fields were completely renamed (e.g. "t_high_water_threshold" to "high_water_arrival_hour" or "t_normal" to "below_bank_return_hour").

NOTE: The SQL was all tested to be functional, but none of the MAPX files have been tested. I figured it's technically easiest to wait until this PR is officially merged into TI and deployed before testing and verifying all of the MAPXs publish the services as expected. If we decide that a more thorough the testing/QA/QC is worth being done before merging this PR, I will reluctantly make that effort.