Viz EC2 Services to Lambda - Part 3a - Anomaly (#582)
This PR migrates the Anomaly product from the Viz EC2 to the Viz Python
Preprocessing lambda function. It broadly encapsulates the following
changes:

Changes:
- New anomaly product script in the python_preprocessing lambda function
that supports both 7day and 14day configurations.
- Removal of anomaly pipeline and product files in the source-aws_loosa
library
- A second python_preprocessing lambda function, which uses the same
source code, but with more RAM (I still need to add logic to choose the
correct function, based on the anomaly config. This will be much easier
to do once we deploy this to TI next week, so I will do it then).

Deployment Considerations:
- I may need to resolve merge conflicts after we merge Part 2 - Rapid
Onset Flooding
- We will need to include a ingest schema db dump when deploying to UAT.

---------

Co-authored-by: CoreyKrewson-NOAA <69687817+CoreyKrewson-NOAA@users.noreply.github.com>
Co-authored-by: NickChadwick-NOAA <nick.chadwick@noaa.gov>