Cached FIM - Part 0 - Terraform and RDS Bastion Setup (#599)
This PR is the first step to implementing the new Cached FIM Workflow
I've been developing (I'll have 1-3 follow-up PRs - I'm trying to split
this all up into appropriately sized chunks).

**This PR:**
- Adds a new IAM role for the Redshift cluster (which notably has
permission to read the viz RDS user secret for external schema setup).
- Adds a new security group for the Redshift cluster
- Adds a Redshift cluster (including a cluster subnet group)
- Adds two steps to the RDS Bastion:
- Redshift cluster setup: Adds a user for use in the pipelines & sets up
a external schema pointed to the ingest schema on viz RDS database (I
didn't abstract this yet, like what was done for the foreign data
wrapper connections on RDS, since it's only one schema, but we could do
that in the future)
- DB_Link Foreign Data Wrapper From RDS to Redshift: I went ahead and
put this in a separate util script, as the syntax is a little different
than the other foreign data wrapper connections between RDS instances
(it's possible that the same syntax could work for all, but I'm not
prioritizing figuring that our right now).
  
**Testing / Deployment Considerations:**
I've tested the IAM Role and Redshift setup in terraform, and the RDS
Bastion steps manually on both Redshift and RDS... but I haven't tested
it all together through the central terraform module yet - I'm hoping
this should be pretty smooth, and I can help troubleshoot any deployment
issues in real-time, Nick. I've also included a couple comments in the
code below for your consideration.