RNR Hotfix: Workflow modified to save and reassign status rather than recalculate (#603)
The RFC 5-Day Max Downstream Streamflow service was completely
recalculating its status based on comparing the output RnR flows to
streamflow thresholds where available. Because streamflow thresholds are
not available for every site, certain sites were not able to be
categorized. This is silly since the flood status of every relevant
point is already pre-calculated as the starting point for the RnR
WRF-Hydro inputs. Thus, this fix writes those statuses to the db during
pre-processing and then joins to that table to re-assign the status at
post-processing.

These changes fix a number of issues reported by WPOD:
https://vlab.noaa.gov/redmine/issues/124216
https://vlab.noaa.gov/redmine/issues/124212
https://vlab.noaa.gov/redmine/issues/124202
https://vlab.noaa.gov/redmine/issues/124165

The changes herein were hotfixed in-place, manually, on Production.