Adds script for fulfilling schism data requests (#610)
All is pretty much in the title. In the past 6 months or so I have
received two requests for schism (coastal) FIM data. I was foolish and
didn't formalize the work I had done on the first request, so when this
second one came in (see ticket #609), I made sure to save the work. So
here it is.

This script could be made more generic-user friendly, but as it stands
now, a user would only need to modify lines 546 through 564 with the
specifics per the request.