Updates rnr to rely upon rfc_max_forecast service (#626)
The rfc_max_forecast service was just re-designed to rely directly upon
the WRDS database rather than the API. This came with some value-added
updates and bug-fixes. With those in place, it made sense to also
re-design the Replace and Route workflow to now rely fully upon the
rfc_max_forecast service data rather than essentially recreating that
foundation itself.

This completely aligns the RnR domain with the RFC Max Forecast domain -
although the RnR domain will still be a subset of that. But before,
there was a chance for a point being in RnR that was not in RFC Max
Forecast in the case of flow-only forecasts.

A new column, rating_curve_source, is also added to the
rfc_based_5day_max_forecast service, which will indicate which rating
curve, if any, was used to produce the flows at each forecast point.