ANA Anomaly Lambda Storage Fix (#627)
This is a super simple bug fix that changes the way the ANA Anomaly
loops through files in the python_preprocessing lambda function.
Downloaded NWM files are now deleted as data is loaded into a dataframe
in memory, so that ephemeral storage does not run out. RAM still seems
stable at about ~7GB used (of 10).

This was done because recent heavier weather was causing this function
to fail due to no space left on device.