Fixes rfc_max_forecast issue where view did not exist by creating the view on deploy (#635)
After deploying the latest uat branch to UAT, the rfc_max_forecast
service was failing to execute because the rnr.flow_thresholds and
rnr.stage_thresholds views did not exist. I had created them manually
during testing and in my short-sidedness excluded the code that would
ensure they get created across new environments on deploy. Here it is
now.

Note: I really hope that the logic I added in
[Core/EC2/RDSBastion/scripts/utils/setup_foreign_tables.tftpl](https://github.com/NOAA-OWP/hydrovis/compare/uat...view-missing-bugfix?expand=1#diff-9ffe547bbfad25c6c32886f6630d8b9458a4ed64d141fce34ab7d347affda781)
is correct. I was not able to test it, but did all the due diligence I
could to ensure it is sound.