Fixes bug with viz_db_ingest Lambda (#652)
I found a bug while attempting to execute an archive case, and while
fixing it also noticed another messy section of code that I cleaned up.

The bug was that if a CSV file failed to upload, it would supposedly
work on the retry, but turns out it wasn't actually working, but failing
silently. This is because the end of the in-memory file object was being
reached just prior to failing, and so on the retry, there is nothing
left of the file to load without first doing a `f.seek(0)`.

The messy part of the code was an unnecessary `with` context surrounding
a large block of code in which the actual couple of lines that did need
the `with` context had what was essentially a duplicate context. I
deleted the outer, unnecessary context and thus indented everything back
one level.