Removes "config" field from services
I couldn't figure out why this "config" field was in there in the first place -
there was no trace of it anywhere in the history I dug through. Thus, I'm
confused that these services ever worked, but they certainly did/do, since
they are currently working in production with this field in there. Either way,
this field was always marked with visible: false, so it never would have
actually showed up anyway...